Fix corruption of PATH env variable when using auto activation mode
(conda-env-activate-path "/") for example
(which is invoked when switching to buffers not visiting any file,
such as magit commit buffers called from emacsclient)
causes resetting of PATH variable to "".  On MacOS, only calling
exec-path-from-shell-initialize can remedy the situation, which makes
in-between emacs unusable (git for magit/vc, ls for dired ... are not
anymore accessible).